Significant refactor to change the approach for finding and creating bff for all dependent projects
Previously, the dependencies for a project were found recursively and added to a set. Dependent projects were tracked and certain properties were set on dependent projects.
Now, bff are created as the recursive loop unrolls. Dependencies are stored and properties are references from them. A map is used to find previously evaluated projects.